fix: handle multiple pictures on map posts
